pyproject.toml
[tool.poetry]
name = "susanomusicbot"
version = "1.2.12"
description = ""
authors = ["alessandrogia <alexgiacento@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
wavelink = "^3.4.1"
python-dotenv = "^1.0.1"
pynacl = "^1.5.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
pytest = "^8.3.3"

[tool.pylint]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
